perf(cfg): shrink Event struct via EventType u8 backing
EventType had u32 backing (4 bytes for 7 variants).  Narrow to u8 — struct
layout now 1+1+pad+4+4+4 = 16 bytes instead of 20.

Saves 4 bytes × ~300K events = ~1.2 MB memory traffic per resolveFull on
typescript.js.  js_buffer's @intFromEnum casts still compile cleanly (u8 → u32
widening is implicit).

+++ src/parser/code_path.zig
@@ -75,7 +75,7 @@ pub const CodePath = struct {
 
 // ── Event ────────────────────────────────────────────────────────
 
-pub const EventType = enum(u32) {
+pub const EventType = enum(u8) {
     codepath_start = 0,
     codepath_end = 1,
     seg_start = 2,
